requirements.txt
pip~=23.1.2
wheel==0.40.0
varint~=1.0.2
termcolor~=2.3.0
sniffio~=1.3.0
setuptools==67.7.2
six~=1.16.0
h11~=0.14.0
idna~=3.4
ecdsa~=0.18.0
netaddr~=0.8.0
multiaddr~=0.0.9
httpcore~=0.17.0
anyio~=3.6.2
base58~=2.1.1
certifi~=2022.12.7
asgiref~=3.6.0
appdirs~=1.4.4
charset-normalizer~=3.1.0
pickleDB~=0.9.2
mnemonic~=0.20
hdwallet~=2.2.1
requests~=2.29.0
httpx~=0.24.0
Django~=4.2
python-dotenv~=1.0.0
pycryptodome~=3.17
IPFS-Toolkit~=0.5.6